LPS-108484 Move useSyncValue to its own folder
