+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-field-type/src/main/resources/META-INF/resources/Text/Text.es.js
@@ -21,38 +21,10 @@ import {normalizeFieldName} from 'dynamic-data-mapping-form-renderer';
 import React, {useEffect, useMemo, useRef, useState} from 'react';
 
 import {FieldBaseProxy} from '../FieldBase/ReactFieldBase.es';
+import {useSyncValue} from '../hooks/useSyncValue';
 import getConnectedReactComponentAdapter from '../util/ReactComponentAdapter.es';
 import {connectStore} from '../util/connectStore.es';
 import templates from './TextAdapter.soy';
-
-/**
- * Use Sync Value to synchronize the initial value with the current internal
- * value, only update the internal value with the new initial value if the
- * values are different and when the value is not changed for more than ms.
- */
-const useSyncValue = newValue => {
-	// Maintains the reference of the last value to check in later renderings if the
-	// value is new or keeps the same, it covers cases where the value typed by
-	// the user is sent to LayoutProvider but it does not descend with the new changes.
-	const previousValueRef = useRef(newValue);
-
-	const [value, setValue] = useState(newValue);
-
-	useEffect(() => {
-		const handler = setTimeout(() => {
-			if (value !== newValue && previousValueRef.current !== newValue) {
-				previousValueRef.current = newValue;
-				setValue(newValue);
-			}
-		}, 300);
-
-		return () => {
-			clearTimeout(handler);
-		};
-	}, [newValue, value]);
-
-	return [value, setValue];
-};
 
 const Text = ({
 	disabled,
